Rename quasar.sbtdatasource to quasar.sbt
